+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Master
+* Improve stream format error [#950](https://github.com/membraneframework/membrane_core/pull/950)
+
 ## 1.2.0
  * Add `:max_instances` option for dynamic pads. [#876](https://github.com/membraneframework/membrane_core/pull/876)
  * Implement `Membrane.Connector`. [#904](https://github.com/membraneframework/membrane_core/pull/904)