+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Master
+## 1.2.1
  * Improve stream format error. [#950](https://github.com/membraneframework/membrane_core/pull/950)
  * Minor fixes in `Membrane.Connector`. [#952](https://github.com/membraneframework/membrane_core/pull/952)